test: fix naming + remove unused code
